Move classic-mod package loaders and support files to Mods.Cnc.
