+++ OpenRA.Mods.Cnc/FileSystem/MixFile.cs
@@ -14,9 +14,12 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using OpenRA.FileFormats;
+using OpenRA.FileSystem;
+using OpenRA.Mods.Cnc.FileFormats;
 using OpenRA.Primitives;
+using FS = OpenRA.FileSystem.FileSystem;
 
-namespace OpenRA.FileSystem
+namespace OpenRA.Mods.Cnc.FileSystem
 {
 	public class MixLoader : IPackageLoader
 	{
@@ -222,7 +225,7 @@ namespace OpenRA.FileSystem
 				return index.ContainsKey(filename);
 			}
 
-			public IReadOnlyPackage OpenPackage(string filename, FileSystem context)
+			public IReadOnlyPackage OpenPackage(string filename, FS context)
 			{
 				IReadOnlyPackage package;
 				var childStream = GetStream(filename);
@@ -242,7 +245,7 @@ namespace OpenRA.FileSystem
 			}
 		}
 
-		bool IPackageLoader.TryParsePackage(Stream s, string filename, FileSystem context, out IReadOnlyPackage package)
+		bool IPackageLoader.TryParsePackage(Stream s, string filename, FS context, out IReadOnlyPackage package)
 		{
 			if (!filename.EndsWith(".mix", StringComparison.InvariantCultureIgnoreCase))
 			{
